requirements.txt
requests
python-telegram-bot
yaml
